1st Project Jet Brains Academy
About:
If you are a beginner web developer, you need to know the basics of CSS and HTML. This project is great for those who need somewhere to start. Your goal is to create a stylized page with flashcards, each containing a question on one side and the answer on the other. You can choose any subject you want for your flashcards!